Fixed code

+++ src/Core/Markets/Bybit/Handlers/Websocket/BybitKlineStreamHandler.cpp
@@ -18,16 +18,16 @@ void Core::Markets::BybitKlineStreamHandler::handle(const QJsonObject &obj, IMar
     {
         QJsonObject data = val.toObject();
         kline.m_category = Tools::stringToMarketType(streamer->id());
-        kline.m_open = data["open"].toString().toDouble();
-        kline.m_close = data["close"].toString().toDouble();
-        kline.m_high = data["high"].toString().toDouble();
-        kline.m_low = data["low"].toString().toDouble();
+        kline.m_open = data["open"].toString();
+        kline.m_close = data["close"].toString();
+        kline.m_high = data["high"].toString();
+        kline.m_low = data["low"].toString();
         kline.m_confirm = data["confirm"].toBool();
         kline.m_start = data["start"].toVariant().toLongLong();
         kline.m_end = data["end"].toVariant().toLongLong();
         kline.m_interval = Tools::stringToInterval(data["interval"].toString());
-        kline.m_volume = data["volume"].toString().toDouble();
-        kline.m_turnover = data["turnover"].toString().toDouble();
+        kline.m_volume = data["volume"].toString();
+        kline.m_turnover = data["turnover"].toString();
     }
 
     emit streamer->klineUpdated(kline);
